refactor: added TritonServerServiceAbs abstract class
